requirements.txt
# This project requires Python 3.10 or higher
Django>=5.2.1
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.9
google-generativeai>=0.3.1
openai>=1.6.1
python-dotenv>=1.0.0
requests>=2.32.0
numpy>=2.2.0
gunicorn>=21.2.0
markdown>=3.8 